Refactors theme events
