+++ src/Domain/Entities/Themes/ThemeEquipmentSlot.cs
@@ -9,9 +9,5 @@ public enum ThemeEquipmentSlot
     Leg,
     MountHarness,
     Mount,
-    Weapon0,
-    Weapon1,
-    Weapon2,
-    Weapon3,
-    WeaponExtra,
+    Weapon,
 }
